Move test files to compiled module for S1479
