+++ java-checks-test-sources/src/main/java/checks/SwitchWithTooManyCasesCheckCustom.java
@@ -1,92 +1,97 @@
-class A{
+package checks;
+
+class SwitchWithTooManyCasesCheckCustom {
   void foo() {
-    switch (1) { // Noncompliant {{Reduce the number of non-empty switch cases from 35 to at most 5.}}
+    switch (1) { // Noncompliant [[sc=5;ec=11]] {{Reduce the number of non-empty switch cases from 35 to at most 5.}}
       case 1:
         System.out.println("");
-      case 1:
+      case 2:
         break;
-      case 1:
+      case 3:
         break;
-      case 1:
+      case 4:
         break;
-      case 1:
+      case 5:
         break;
-      case 1:
+      case 6:
         break;
-      case 1:
+      case 7:
         break;
-      case 1:
+      case 8:
         break;
-      case 1:
+      case 9:
         break;
-      case 1:
+      case 10:
         break;
-      case 1:
+      case 11:
         break;
-      case 1:
-        break;
+      case 12:
         System.out.println("");
-      case 1:
         break;
-      case 1:
+      case 13:
         break;
-      case 1:
+      case 14:
         break;
-      case 1:
+      case 15:
         break;
-      case 1:
+      case 16:
         break;
-      case 1:
+      case 17:
         break;
-      case 1:
+      case 18:
         break;
-      case 1:
+      case 19:
         break;
-      case 1:
+      case 20:
         break;
-      case 1:
+      case 21:
         break;
-      case 1:
+      case 22:
         break;
-      case 1:
+      case 23:
         break;
-      case 1:
+      case 24:
         break;
-      case 1:
+      case 25:
         break;
-      case 1:
+      case 26:
         break;
-      case 1:
+      case 27:
         break;
-      case 1:
+      case 28:
         break;
-      case 1:
+      case 29:
         break;
-      case 1:
+      case 30:
         break;
-      case 1:
+      case 31:
         break;
-      case 1:
+      case 32:
         break;
-      case 1:
+      case 33:
         break;
+      case 34:
+      case 35:
+      case 36:
+      case 37:
         System.out.println("");
+        break;
       default:
         System.out.println("");
     }
-    switch (1) { // Noncompliant [[secondary=78,80,82,84,86,88]]
+    switch (1) { // Noncompliant [[secondary=+1,+3,+5,+7,+9,+11]]
       case 1:
         break;
-      case 1:
+      case 2:
         break;
-      case 1:
+      case 3:
         System.out.println("");
-      case 1:
+      case 4:
         break;
-      case 1:
+      case 5:
         break;
-      case 1:
-      case 1:
+      case 6:
+      case 7:
         System.out.println("");
     }
   }
