+++ java-checks/src/test/java/org/sonar/java/checks/SwitchWithTooManyCasesCheckTest.java
@@ -22,12 +22,14 @@ package org.sonar.java.checks;
 import org.junit.jupiter.api.Test;
 import org.sonar.java.checks.verifier.JavaCheckVerifier;
 
+import static org.sonar.java.checks.verifier.TestUtils.testSourcesPath;
+
 class SwitchWithTooManyCasesCheckTest {
 
   @Test
   void defaultValue() {
     JavaCheckVerifier.newVerifier()
-      .onFile("src/test/files/checks/SwitchWithTooManyCasesCheck.java")
+      .onFile(testSourcesPath("checks/SwitchWithTooManyCasesCheck.java"))
       .withCheck(new SwitchWithTooManyCasesCheck())
       .verifyIssues();
   }
@@ -37,7 +39,7 @@ class SwitchWithTooManyCasesCheckTest {
     SwitchWithTooManyCasesCheck check = new SwitchWithTooManyCasesCheck();
     check.maximumCases = 5;
     JavaCheckVerifier.newVerifier()
-      .onFile("src/test/files/checks/SwitchWithTooManyCasesCheckCustom.java")
+      .onFile(testSourcesPath("checks/SwitchWithTooManyCasesCheckCustom.java"))
       .withCheck(check)
       .verifyIssues();
   }
